pyproject.toml
[project]
name = "manga-scraper"
version = "1.0.0"
description = "Production-ready enterprise manga scraper with Cloudflare bypass"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    
    # Database
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    
    # Task Queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "flower>=2.0.1",
    
    # HTTP Client
    "httpx>=0.28.0",
    "aiohttp>=3.11.0",
    
    # Browser Automation (Cloudflare bypass)
    "nodriver>=0.38.0",
    
    # HTML Parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    
    # Image Processing
    "pillow>=11.0.0",
    
    # Object Storage
    "boto3>=1.35.0",
    "aioboto3>=13.0.0",
    
    # Utilities
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.1",
    "orjson>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
